+++ crates/aranya-policy-lang/src/lang/parse/markdown.rs
@@ -7,7 +7,7 @@ use markdown::{
 };
 use serde::Deserialize;
 
-use super::{ParseError, ParseErrorKind, Version, parse_policy_chunk};
+use super::{ParseError, ParseErrorKind, Version, error, parse_policy_chunk};
 
 #[derive(Deserialize)]
 struct FrontMatter {
@@ -22,11 +22,11 @@ fn parse_front_matter(yaml: &Yaml) -> Result<Version, ParseError> {
         "2" => Version::V2,
         v => {
             return Err(ParseError::new(
-                ParseErrorKind::InvalidVersion {
+                error::InvalidVersion {
                     found: v.to_string(),
                     required: Version::V2,
                 },
-                "Update `policy-version`.".to_string(),
+                Version::help_message(),
                 None,
             ));
         }
@@ -42,8 +42,6 @@ pub struct PolicyChunk {
 
 #[derive(Default, Debug)]
 pub struct ChunkOffset {
-    /// 0-based line offset of policy code within document.
-    pub line: usize,
     /// 0-based byte offset of policy code within document.
     pub byte: usize,
 }
@@ -74,10 +72,6 @@ fn extract_policy_from_markdown(node: &Node) -> Result<(Vec<PolicyChunk>, Versio
             {
                 let point = &c.position.as_ref().expect("no code block position").start;
 
-                // The 1-based start line of the code block is
-                // the 0-based start line of the policy code.
-                let line = point.line;
-
                 // The starting position of the code block is
                 // the triple-backtick, so add three for the
                 // backticks, six for the language tag, and
@@ -89,7 +83,7 @@ fn extract_policy_from_markdown(node: &Node) -> Result<(Vec<PolicyChunk>, Versio
 
                 chunks.push(PolicyChunk {
                     text: c.value.clone(),
-                    start: ChunkOffset { line, byte },
+                    start: ChunkOffset { byte },
                 });
             }
         }